Moves AdminRoute to its own file
